+++ resto-code-front/src/Components/HEADER/Header.css
@@ -34,7 +34,7 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  width: 12.5rem;
+  width: fit-content;
   font-family: "Bebas Neue", sans-serif;
 }
 .headerModal {
